Squashed commit of the following:
commit c0ffcc0bc8602386b484effa2ea371177dbf8247
Author: Hyungjun Kim <73098596+Rocketstein@users.noreply.github.com>
Date:   Sat Jun 6 14:12:41 2026 +0900

    Feat: Add SoftPaused state to world tick

commit 099b101e5714a39db7bf59e90a78b241a5d809e6
Author: Hyungjun Kim <73098596+Rocketstein@users.noreply.github.com>
Date:   Sat Jun 6 14:06:52 2026 +0900

    Minor update

commit 842e29af65ad180d4ad690c015c19726ec764524
Author: Hyungjun Kim <73098596+Rocketstein@users.noreply.github.com>
Date:   Sat Jun 6 14:02:33 2026 +0900

    Fix: Exclude actors with tag SoftPauseImmune

commit 1fc91d3bff2bff7ae49500030f17112cac2edb1c
Author: Hyungjun Kim <73098596+Rocketstein@users.noreply.github.com>
Date:   Sat Jun 6 13:56:21 2026 +0900

    Update GameSoftPauseState implementation

commit cbf7c7a2ac7f3934f009deabdb3103589229fc01
Author: Hyungjun Kim <73098596+Rocketstein@users.noreply.github.com>
Date:   Sat Jun 6 13:28:22 2026 +0900

    Softpause prototype

+++ KraftonEngine/Source/Game/GameMode/AFinaleGameMode.h
@@ -3,7 +3,7 @@
 
 #include "Source/Game/GameMode/AFinaleGameMode.generated.h"
 
-enum class EGamePhase : uint8;
+enum class EGamePhase : uint16;
 
 UCLASS()
 class AFinaleGameMode : public AGameModeBase
